+++ src/schemapack/_internals/exceptions.py
