+++ src/schemapack/_internals/isolate.py
