+++ src/schemapack/_internals/spec/__init__.py
