+++ src/schemapack/_internals/spec/custom_types.py
