+++ src/schemapack/_internals/spec/datapack.py
