+++ src/schemapack/_internals/utils.py
