+++ src/schemapack/utils.py
@@ -17,6 +17,6 @@
 
 """Utility functions."""
 
-from schemapack.__internals.utils import FrozenDict, read_json_or_yaml_mapping
+from schemapack._internals.utils import FrozenDict, read_json_or_yaml_mapping
 
 __all__ = ["read_json_or_yaml_mapping", "FrozenDict"]
